requirements.txt
librosa
whisper-openai
torch
torchaudio
soundfile
numpy